encapsulate components
